requirements.txt
fastapi
uvicorn
tortoise-orm
aerich
httpx
pydantic-settings
google_auth_oauthlib
aiomysql
openai
python-multipart